Dictionary: Implement JMDict parser.
This commit implements the JMDict parser using SwiftyJSON. The parser receives a JMDict formatted as a JSON and parses most of its entities and attributes into NSManagedObjects defined in the Dictionary directory. This commit also defines an NSManagedObject extension that adds a new convenience initializer that should be used for all NSManagedObjects in order to resolve ambiguous context warnings.