DictionaryTests: comment out reading test for CI
The testReadInDictionaryData function should only exist for developers to see if they have correctly set up their dictionary. The test will always fail in CI because the dictionary file is ommitted in the repo git .gitignore. This commit comments out the function so devs can use it in the future without failing CI.