Reader: Implemented page flipping and word tapping (#7)
This commit uses SwiftUIPager to implement pagination. In order to accurately paginate according to different screen sizes and text fonts, we borrow UITextView from UIKit, which helps us calculate the amount of text that fits into the frame. This commit also implements word tapping. We used UIViewRepresentable to use UITextView's functionality that allows us to get the location of a tap and its corresponding word.