Components: Support adding a search bar to NavigationView.
Enables a SwiftUI NavigationView to include a search bar. We add a UIViewControllerRepresentable that retrieves the UINavigationController of a SwiftUI view component. We also implement a SearchBar class that wraps UISearchController to be instantiated when a Search Bar is needed. Finally, we wrap SearchBar in a ViewModifier and extend View with a method to easily augment NavigationViews with SwiftUI-like syntax.