Discover/TrendingView: Refactor into TrendingView, implement search bar.
This commit refactors the current DiscoverView into a TrendingView group. This commit also implements the UI for the TrendingView's search bar, such that the NavigationView title collapses on scroll and the search bar is pinned to the top regardless of scroll position. The search bar itself also clears current input text and ends editing on cancel.