DiscoverView: Category filter buttons and NavLink changes.
Implements an HStack of category filter buttons at the top of DiscoverView. In order to scroll past this row of buttons, we refactor the DiscoverView's List to use ScrollView + LazyVStack instead. We also change the NavLink for each row to be constrained to the right end of the row. This change is important for enabling novel title definitions in the future. Lastly, we refactor SearchResults view's List to match the architecture of DiscoverView for maintainability.