Shiori/Tokenizer: Implement Tokenizer
Implements the Tokenizer class, which uses the DictionaryFetcher, Deinflector, and MecabWrapper. The MecabWrapper is used for preliminary tokenization, after which Tokenizer attempts to concatenate tokens and deinflect their representations. During this stage, Tokenizer references DictionaryFetcher to determine whether two tokens can be concatenated. As opposed to MecabWrapper, Tokenizer returns Tokens, which will not be modified after they are created, hence the reason they are structs. Tokens also contain their relevant dictionary entry, which is useful for retrieving the token's definition, and the range of indices in the original string that comprise the token, which will be useful when implementing functionaity to tap on a token.