rawger/{NovelDetails, Reader}: Hide tab bar.
This commit imports the SwiftUI Introspect library, which allows us to easily access the underlying UIKit components in SwiftUI. Introspect is then used in the Reader view and NovelDetails to hide the tab bar, which was originally obstructing part of the Definer. Furthermore, with the tab bar hidden, the bottom sheet nature of the Definer now appears to render from the bottom of the screen rather than sprout from the tab bar.