rawger/Reader: Disable scroll on Reader View, fix bottom margin.
This commit addresses and issue where certain pages will become scrollable when the text pushes against the bottom edge of the DefinableTextView. We disable this behavior by having the DefinableTextView take up as much available space as possible, and then constraining its bottom with a placeholder rectangle at the location of the minimized Definer.