rawger/{Reader, SectionNavigation}: Fix two bugs.
Fixes a bug where the current section is not highlighted in SectionNavigationView, and another where the loading page never disappears. Both were caused by ReaderView not having its viewModel as an ObservedObject. Fixing the first bug also required passing a binding to the sectionNcode down the ReaderView hierarchy and into the navigation bar.