hugo/{Reader, Components}: Paginator Optimization
Avoided excessive copying and substringing of attributed string in the paginate function and lengthThatFits function by utilizing CFRanges and a running sum of the length of each page. Fixed a bug where paginator is called multiple times causing paginate to run on each load. This last issue is fixed by changing ObservedObject to StateObject.