hugo/WKText: Implemented drag to reveal progress indicators
Progress indicators are now only visible when users drag from the top of the page or the bottom of the page. Change progress indicator size. Dynamically sized the progress indicator by the offset of the webkit's scroll view, which fixes an issue where there is extra space on top of the contents of WKText.