hugo/{Discover, DefinableText}: Fixed highlighiting issue
Fix an issue where multiple rows would have words highlighted due to the last selected row being deallocated as we are scrolling past it. Fixed by introducing an id to the DefinableTextView and using the id, we could check for equality between the newly allocated rows and the old rows to regain reference to the last selected row.