rawger/AppCentral: Fix phantom tab view
Fixes an issue where the tab view will appear when the content of the screen touches the tab view for certain views, and only on an actual device. We fix this by using introspect to hide the tab view per tab, as opposed to hiding it for the entirety of the tab view.