hugo/DefinableText: Switched order of custom draw and built in draw
Fixed an issue where the highlight is drawn after the text, therefore changing the text color. Switched the order so that highlight is drawn below the text. This is doable because the selected content is attached with a clear background color, so when it is drawn on top of the yellow highlight, it does not obstruct it.