rawger/{WKText, NovelDetails}: Definable NovelDetails title
Makes the title in NovelDetailsView definable by rendering it with a WKText. Since there are two WKTexts, we keep track of a lastSelectedWktextViewModel similar to DefinableText to ensure that only one token is ever highlighted across both views. Also factors out the styling script into its own file and conditionally styles the WKText according to the type of content being rendered.